Rename rc directory to src
