+++ README.md
@@ -9,11 +9,15 @@ npm install
 npm run dev
 ```
 
-Source files are located in the `rc/` directory with the following structure:
+Source files are located in the `src/` directory with the following structure:
 
 ```
-rc/
+src/
   components/
+    common/
+    layout/
+    charts/
+    tables/
   pages/
   hooks/
   context/
