+++ src/App.jsx
@@ -6,8 +6,8 @@ import Market from './pages/Market';
 import Messages from './pages/Messages';
 import Portfolio from './pages/Portfolio';
 import Settings from './pages/Settings';
-import ThemeToggle from './components/ThemeToggle';
-import Sidebar from './components/Sidebar';
+import ThemeToggle from './components/common/ThemeToggle';
+import Sidebar from './components/layout/Sidebar';
 import styles from './App.module.scss';
 
 function App() {
