+++ src/components/charts/ChartLegend.module.scss
