+++ src/components/charts/MetricComparison.jsx
