+++ src/components/common/LoadingIndicator.module.scss
