+++ src/components/tables/TradeHistoryTable.module.scss
